filter chains reworked
ImageResizeDefinition now allows to add filters to the resize filter chain.
The additional filter chain is now called post_filters and is ALWAYS executed
even if the image is not going to be resized. Filters can not be added in the
constructor any more but only using ImageResizeDefinition::addFilter() or
ImageResizeDefinition::addPostFilter()